Sprint13 - Actividad complementaria  (#19)
* Cambio en user.model.js

* Agergado de owner_id

* fixes user model

* fixes comment dto

* fixes product dto

* fixes user dto

* Agregado de user_id en data

* cambio de ARS a USD

* Agregado de Types

* corrección de nombre

* cambio de nombre

* Agregado de owner_id y eliminación de comentarios

* mejoras en código

* Agregado de condicional para modificar un producto

* agregado de user_id al token

* modificación de la respuesta de las politicas

* errores

* modificación de destroy con condicional de rol de usuario

* Agregado de endpoint para cambiar el rol de user a prem

* agregado de enpoint

* Mejoras en código

* Agrego otras politicas a products

* cambios en politicas de read y readone

* eliminación de politica public, siempre se verifica por token

* cambios en politicas

* Agregado de error al diccionario

* Agregado de product service para evaluar el producto agregado

* errors log

* Fixes

* eliminación de politica publica en router

* mejoras en condicionas de roles

* errors log

* Eliminación de poitica PUBLIC

* Mejoras en código

* Cambios de politca PUBLIC

* Fixes

* Fixes

* Fixes products default link

* Fixes

* modificaciones en users

* Creación de mail de recuperación

* RECOVERY SERVICE

* Nuevo elemento en diccionario

* modificaciones en función de utilidad

* fixes

* eliminación de current por no usarse

* modificación de controller user

* res.json

* cambio en response

* agregado de tiempo de expiración

* cambio de link

* cambio en función createToken

* ap -cookies

* envio id por link para usarlo por params en front

* modificación de tiempo de expiración

* fixes

* cambio en función

* cambios para el restore

* Modificación de Token!

* Creación de función para verificar el token

* Nueva ruta para verificación de token

* Nuevo diccionario de errores

* errores

* Fin de actividad

* Eliminación de console.log

* errors logger

+++ AP-Server/src/controllers/products.controller.js
@@ -7,6 +7,7 @@ class ProductsController {
     create = async (req, res, next) => {
         try {
             const data = req.body;
+            data.user_id = req.user._id;
             const response = await this.service.create(data);
             return res.success201(response);
         } catch (error) {
@@ -25,9 +26,15 @@ class ProductsController {
             if (req.query.title) {
                 filter.title = new RegExp(req.query.title.trim(), "i");
             }
+
             if (req.query.sort === "desc") {
                 options.sort.title = "desc";
             }
+
+            if (req.user && req.user.role === "PREM") {
+                filter.owner_id = { $ne: req.user._id };
+            }
+
             const all = await this.service.read({ filter, options });
             return res.success200(all);
         } catch (error) {
@@ -47,6 +54,13 @@ class ProductsController {
         try {
             const { pid } = req.params;
             const data = req.body;
+            if (req.user.role === "PREM") {
+                const one = await this.service.readOne(pid);
+                const oid = one.owner_id.toString();
+                if (oid !== req.user.user_id) {
+                    return res.error403();
+                }
+            }
             const response = await this.service.update(pid, data);
             return res.success200(response);
         } catch (error) {
@@ -56,8 +70,19 @@ class ProductsController {
     destroy = async (req, res, next) => {
         try {
             const { pid } = req.params;
-            const response = await this.service.destroy(pid);
-            return res.success200(response);
+            if (req.user.role === "PREM") {
+                const one = await this.service.readOne(pid);
+                const oid = one.owner_id.toString();
+                if (oid === req.user.user_id) {
+                    const response = await this.service.destroy(pid);
+                    return res.success200(response);
+                } else {
+                    return res.error403();
+                }
+            } else {
+                const response = await this.service.destroy(pid);
+                return res.success200(response);
+            }
         } catch (error) {
             return next(error);
         }
